Refactor context
 - Refactor context.

 - Move last exit from global context to a scope.

 - Rearrange AST items.

 - Fix spelling issues.

 - Rearrange parse crate.
