+++ crates/pjsh_ast/src/lib.rs
@@ -1,88 +1,13 @@
 mod command;
 mod control;
 mod io;
+mod pipeline;
 mod program;
-
-use std::collections::VecDeque;
+mod word;
 
 pub use command::Command;
 pub use control::{ConditionalChain, ConditionalLoop};
-pub use io::{FileDescriptor, Redirect, RedirectOperator};
-pub use program::Program;
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Statement {
-    AndOr(AndOr),
-    Assignment(Assignment),
-    Function(Function),
-    If(ConditionalChain),
-    While(ConditionalLoop),
-    Subshell(Program),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Assignment {
-    pub key: Word,
-    pub value: Word,
-}
-
-impl Assignment {
-    pub fn new(key: Word, value: Word) -> Self {
-        Self { key, value }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Function {
-    pub name: String,
-    pub args: VecDeque<String>,
-    pub body: Program,
-}
-
-impl Function {
-    pub fn new(name: String, args: VecDeque<String>, body: Program) -> Self {
-        Self { name, args, body }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Word {
-    Literal(String),
-    Quoted(String),
-    Variable(String),
-    Subshell(Program),
-    ProcessSubstutution(Program),
-    Interpolation(Vec<InterpolationUnit>),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum InterpolationUnit {
-    Literal(String),
-    Unicode(char),
-    Variable(String),
-    Subshell(Program),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct AndOr {
-    pub operators: Vec<AndOrOp>,
-    pub pipelines: Vec<Pipeline>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum AndOrOp {
-    And,
-    Or,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Pipeline {
-    pub is_async: bool,
-    pub segments: Vec<PipelineSegment>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum PipelineSegment {
-    Command(Command),
-    Condition(Vec<Word>),
-}
+pub use io::{FileDescriptor, Redirect, RedirectMode};
+pub use pipeline::{Pipeline, PipelineSegment};
+pub use program::{AndOr, AndOrOp, Assignment, Function, Program, Statement};
+pub use word::{InterpolationUnit, Word};
