+++ crates/pjsh_core/src/lib.rs
@@ -1,13 +1,10 @@
 pub mod command;
 mod condition;
-mod context;
 mod env;
 mod fs;
 pub mod utils;
 
 pub use condition::Condition;
-// pub use old_context::Context;
-pub use context::{Context, Scope};
-pub use env::host::Host;
 pub use env::std_host::StdHost;
+pub use env::{context::Context, context::Scope, host::Host};
 pub use fs::{find_in_path, paths};
