+++ crates/pjsh_exec/src/tests/and_or.rs
@@ -28,7 +28,7 @@ fn execute_and_success() {
         pipelines: vec![pipeline(0), pipeline(0)],
     };
     executor.execute_and_or(and_success, Arc::clone(&ctx), &fds);
-    assert_eq!(ctx.lock().last_exit, 0);
+    assert_eq!(ctx.lock().last_exit(), 0);
 }
 
 #[test]
@@ -41,7 +41,7 @@ fn execute_and_fail() {
         pipelines: vec![pipeline(1), pipeline(0)],
     };
     executor.execute_and_or(and_success, Arc::clone(&ctx), &fds);
-    assert_eq!(ctx.lock().last_exit, 1);
+    assert_eq!(ctx.lock().last_exit(), 1);
 }
 
 #[test]
@@ -54,7 +54,7 @@ fn execute_or_first_success() {
         pipelines: vec![pipeline(0), pipeline(1)],
     };
     executor.execute_and_or(and_success, Arc::clone(&ctx), &fds);
-    assert_eq!(ctx.lock().last_exit, 0);
+    assert_eq!(ctx.lock().last_exit(), 0);
 }
 
 #[test]
@@ -67,7 +67,7 @@ fn execute_or_last_success() {
         pipelines: vec![pipeline(1), pipeline(0)],
     };
     executor.execute_and_or(and_success, Arc::clone(&ctx), &fds);
-    assert_eq!(ctx.lock().last_exit, 0);
+    assert_eq!(ctx.lock().last_exit(), 0);
 }
 
 #[test]
@@ -80,5 +80,5 @@ fn execute_or_last_fail() {
         pipelines: vec![pipeline(1), pipeline(1)],
     };
     executor.execute_and_or(and_success, Arc::clone(&ctx), &fds);
-    assert_eq!(ctx.lock().last_exit, 1);
+    assert_eq!(ctx.lock().last_exit(), 1);
 }
