+++ crates/pjsh_parse/Cargo.toml
@@ -5,6 +5,4 @@ license = "MIT"
 edition = "2021"
 
 [dependencies]
-unicode-segmentation = "1.8.0"
-
 pjsh_ast = { path = "../pjsh_ast" }
