+++ crates/pjsh_parse/src/lib.rs
@@ -1,8 +1,8 @@
 mod error;
 mod lex;
 mod parse;
-mod tokens;
+mod token;
 
 pub use error::ParseError;
-pub use lex::lexer::{lex, lex_interpolation};
+pub use lex::{input::Span, lexer::{lex, lex_interpolation}};
 pub use parse::parser::{parse, parse_interpolation};
