+++ crates/pjsh_parse/src/parse/cursor.rs
@@ -1,8 +1,8 @@
 use std::{iter::Peekable, vec::IntoIter};
 
 use crate::{
-    lex::lexer::{Span, Token},
-    tokens::TokenContents,
+    token::{Token, TokenContents},
+    Span,
 };
 
 /// The newline mode determines how a [`TokenCursor`] handles newline tokens.
@@ -57,7 +57,7 @@ impl TokenCursor {
     /// If `func` returns `true` for the next token, consume and return it.
     /// Otherwise, return `None`.
     ///
-    /// Skips trival tokens before evaluating the condition.
+    /// Skips trivial tokens before evaluating the condition.
     pub fn next_if(&mut self, func: impl FnOnce(&Token) -> bool) -> Option<Token> {
         self.skip_trivial_tokens();
         self.tokens.next_if(func)
@@ -66,7 +66,7 @@ impl TokenCursor {
     /// Consume and return the next token if `contents` match the next token's contents.
     /// Otherwise, return `None`.
     ///
-    /// Skips trival tokens.
+    /// Skips trivial tokens.
     pub fn next_if_eq(&mut self, contents: TokenContents) -> Option<Token> {
         self.skip_trivial_tokens();
         self.next_if(|token| token.contents == contents)
