+++ crates/pjsh_parse/src/token.rs
@@ -1,5 +1,23 @@
-use crate::lex::lexer::Token;
+use crate::Span;
 
+/// A unit of input identified through lexical analysis.
+#[derive(Debug, Clone, PartialEq)]
+pub struct Token {
+    /// Token contents.
+    pub contents: TokenContents,
+
+    /// Token position in the input.
+    pub span: Span,
+}
+
+impl Token {
+    /// Constructs a new token.
+    pub fn new(contents: TokenContents, span: Span) -> Self {
+        Self { contents, span }
+    }
+}
+
+/// The contents of a token.
 #[derive(Debug, Clone, PartialEq)]
 pub enum TokenContents {
     /// "# ..."
@@ -71,10 +89,18 @@ pub enum TokenContents {
     Unknown,
 }
 
+/// An interpolation unit within an interpolation token.
 #[derive(Debug, Clone, PartialEq)]
 pub enum InterpolationUnit {
+    /// A literal unit.
     Literal(String),
+
+    /// A unicode character.
     Unicode(char),
+
+    /// The name of a variable unit that is evaluated at runtime.
     Variable(String),
+
+    /// A subshell that is evaluated at runtime.
     Subshell(Vec<Token>),
 }
